+++ Server.Library/Models/ReviewModel.cs
@@ -19,5 +19,5 @@ public class ReviewModel
 
     [Required]
     [Range(0, 5)]
-    public double Rating { get; set; }
+    public int Rating { get; set; }
 }
